Add files via upload
This is the code for my Engineering 1 Capstone project. The project aimed to tell temperature in a creative way. We did this by having an LED that would change its color dynamically based on the temperature of the room along with a motor turning a dial in order to show the user how hot or cold it was in the room. The temperature readings were acquired using a DHT11 temperature and humidity sensor.